Make the download page more "symmetric"
at least try.

Both the Stable and the Development versions now look similar. They
are contain similar artifact, do a degree ;-). The Stable versions
contains the tar ball built by `make dist` and the Development
versions contain a tar ball that is just a tared up source tree. Not
quite the same thing but labeled the same in this case.

Hope the introduction text makes the differences clear.